Jumping into the future

+++ Engine/source/materials/matInstance.cpp
@@ -35,6 +35,7 @@
 #include "gfx/sim/cubemapData.h"
 #include "gfx/gfxCubemap.h"
 #include "core/util/safeDelete.h"
+#include "ts/tsShape.h"
 
 class MatInstParameters;
 
@@ -242,14 +243,16 @@ void MatInstance::construct()
 {
    mUserObject = NULL;
    mCurPass = -1;
-   mProcessedMaterial = false;
+   mProcessedMaterial = NULL;
    mVertexFormat = NULL;
    mMaxStages = 1;
    mActiveParameters = NULL;
    mDefaultParameters = NULL;
    mHasNormalMaps = false;
+   mUsesHardwareSkinning = false;
    mIsForwardLit = false;
    mIsValid = false;
+   mIsHardwareSkinned = false;
 
    MATMGR->_track(this);
 }
@@ -290,6 +293,12 @@ bool MatInstance::init( const FeatureSet &features,
 //----------------------------------------------------------------------------
 bool MatInstance::reInit()
 {
+   if (!mVertexFormat)
+   {
+      mIsValid = false;
+      return mIsValid;
+   }
+
    SAFE_DELETE(mProcessedMaterial);
    deleteAllHooks();
    mIsValid = processMaterial();
@@ -354,6 +363,11 @@ bool MatInstance::processMaterial()
 
       FeatureSet features( mFeatureList );
       features.exclude( MATMGR->getExclusionFeatures() );
+
+      if (mVertexFormat->hasBlendIndices() && TSShape::smUseHardwareSkinning)
+      {
+         features.addFeature( MFT_HardwareSkinning );
+      }
       
       if( !mProcessedMaterial->init(features, mVertexFormat, mFeaturesDelegate) )
       {
@@ -367,10 +381,13 @@ bool MatInstance::processMaterial()
 
       const FeatureSet &finalFeatures = mProcessedMaterial->getFeatures();
       mHasNormalMaps = finalFeatures.hasFeature( MFT_NormalMap );
+      mUsesHardwareSkinning = finalFeatures.hasFeature( MFT_HardwareSkinning );
 
       mIsForwardLit =   (  custMat && custMat->mForwardLit ) || 
                         (  !finalFeatures.hasFeature( MFT_IsEmissive ) &&
                            finalFeatures.hasFeature( MFT_ForwardShading ) );
+
+      mIsHardwareSkinned = finalFeatures.hasFeature( MFT_HardwareSkinning );
 
       return true;
    }
@@ -449,6 +466,12 @@ void MatInstance::setTransforms(const MatrixSet &matrixSet, SceneRenderState *st
    mProcessedMaterial->setTransforms(matrixSet, state, getCurPass());
 }
 
+void MatInstance::setNodeTransforms(const MatrixF *address, const U32 numTransforms)
+{
+   PROFILE_SCOPE(MatInstance_setNodeTransforms);
+   mProcessedMaterial->setNodeTransforms(address, numTransforms, getCurPass());
+}
+
 void MatInstance::setSceneInfo(SceneRenderState * state, const SceneData& sgData)
 {
    PROFILE_SCOPE(MatInstance_setSceneInfo);
@@ -493,6 +516,14 @@ bool MatInstance::hasGlow()
 { 
    if( mProcessedMaterial )
       return mProcessedMaterial->hasGlow(); 
+   else
+      return false;
+}
+
+bool MatInstance::hasAccumulation() 
+{ 
+   if( mProcessedMaterial )
+      return mProcessedMaterial->hasAccumulation(); 
    else
       return false;
 }
